latin_corpus/requirements.txt
pandas
python-Levenshtein
unidecode
rapidfuzz
